pyproject.toml
[project]
name = "ibkr-to-irpf"
version = "0.1.0"
description = "Convert IBKR statements to Brazilian IRPF declaration format"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.123.9",
    "jinja2>=3.1.6",
    "python-bcb>=0.1.8",
    "pytest>=9.0.1",
    "ruff>=0.14.8",
]

[tool.pytest]
python_files = ["tests.py"]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
preview = true
select = [
    "B",    # flake8-bugbear
    "C",    # flake8-comprehensions
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "FURB", # refurb
    "I",    # isort
    "PTH",  # flake8-use-pathlib
    "UP",   # pyupgrade
    "W",    # pycodestyle warnings
]
